MVC pattern
