+++ routes/routes.js
@@ -1,0 +1,18 @@
+const express = require('express');
+const router = express.Router();
+const {
+    getAllTasks,
+    getTaskById,
+    createTask,
+    updateTask,
+    deleteTask,
+} = require('../controllers/taskController');
+
+
+router.get('/', getAllTasks);
+router.get('/:id', getTaskById);
+router.post('/', createTask);
+router.put('/:id', updateTask);
+router.delete('/:id', deleteTask);
+
+module.exports = router;
